# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Slogr Controller API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/add-session": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Create a session",
                "description": "Create a new session",
                "operationId": "3b5e31aacf66d964856cd93ab0127fee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "serve": {
                                        "description": "Server ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "client": {
                                        "description": "Client ID",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "profile": {
                                        "description": "Profile ID",
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "schedule": {
                                        "description": "Schedule",
                                        "type": "integer",
                                        "example": "180"
                                    },
                                    "count": {
                                        "description": "Count",
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "n_packets": {
                                        "description": "Number of packets",
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "p_interval": {
                                        "description": "Packet interval",
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "w_time": {
                                        "description": "Wait time",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "dscp": {
                                        "description": "DSCP",
                                        "type": "integer",
                                        "example": "46"
                                    },
                                    "p_size": {
                                        "description": "Packet size",
                                        "type": "integer",
                                        "example": "1500"
                                    },
                                    "edit": {
                                        "description": "Edit flag",
                                        "type": "boolean",
                                        "example": "false"
                                    },
                                    "aid": {
                                        "description": "Session ID (required if edit flag is true)",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "properties": {
                                                "session": {}
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "300": {
                        "description": "Validation error or client unreachable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete-session": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Delete a session",
                "description": "Delete a session by ID",
                "operationId": "952d37d2a140d3e84cf907bf05c944fa",
                "requestBody": {
                    "description": "Session ID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "delete": {
                                        "description": "ID of the session to be deleted",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the deletion operation",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "security": {
                "passport": {
                    "type": "oauth2",
                    "description": "Laravel passport oauth2 security.",
                    "in": "header",
                    "scheme": "https",
                    "flows": {
                        "password": {
                            "authorizationUrl": "localhost/slogr/oauth/authorize",
                            "tokenUrl": "localhost/slogr/oauth/token",
                            "refreshUrl": "localhost/slogr/token/refresh",
                            "scopes": []
                        }
                    }
                }
            }
        }
    }
}